Move undirected al test entirely into GraphSpec
