+++ src/Gliph/Exception/IncompatibleGraphTypeException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Gliph\Exception;
+
+/**
+ * Exception thrown when an incompatible graph type is provided to an operation.
+ */
+class IncompatibleGraphTypeException extends \LogicException {}
