+++ tests/Gliph/Graph/TestTraits/DirectedGraphSpec.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Gliph\Graph\TestTraits;
+
+/**
+ * Provides a trait to test the methods of DirectedGraph and MutableDirectedGraph.
+ */
+trait DirectedGraphSpec {
+
+}
